+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+# Next Release
+
 # Release v1.5.0
 
 ## Highlights